feat: implement PKCE login flow and refactor authentication components
