+++ package.json
@@ -20,6 +20,7 @@
     "@tailwindcss/vite": "^4.1.13",
     "@tanstack/react-query": "^5.87.1",
     "@tanstack/react-router": "^1.131.36",
+    "@tanstack/zod-adapter": "^1.131.41",
     "axios": "^1.7.8",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -37,7 +38,8 @@
     "styled-components": "^6.1.13",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
-    "universal-cookie": "^7.2.2"
+    "universal-cookie": "^7.2.2",
+    "zod": "^4.1.8"
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
