+++ src/services/authAPI.ts
@@ -8,3 +8,14 @@ export interface LoginRequest {
 export const login = async (dto: LoginRequest) => {
     return post("/api/auth/login", dto);
 }
+
+export interface LoginPkceRequest {
+    method: string;
+    code: string;
+    codeChallenge: string;
+    codeChallengeMethod: string;
+}
+
+export const loginPkce = async (dto: LoginPkceRequest) => {
+    return post("/api/auth/login-pkce", dto);
+}
